refactor(core-negative): split DER helpers into microcrate
