+++ crates/uselesskey-core-negative-der/Cargo.toml
@@ -1,23 +1,22 @@
 [package]
-name = "uselesskey-core-negative"
+name = "uselesskey-core-negative-der"
 version = "0.3.0"
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "Compatibility facade for negative fixture builders used by uselesskey."
+description = "DER-focused negative fixture builders for deterministic corruption in uselesskey."
 readme = "README.md"
 exclude = ["fuzz/**", "corpus/**", "**/*.der", "**/*.pem"]
 categories.workspace = true
-keywords = ["fixtures", "negative", "pem", "der", "testing"]
+keywords = ["fixtures", "negative", "der", "testing", "deterministic"]
 homepage.workspace = true
 authors.workspace = true
 
-documentation = "https://docs.rs/uselesskey-core-negative"
+documentation = "https://docs.rs/uselesskey-core-negative-der"
 
 [dependencies]
-uselesskey-core-negative-der = { workspace = true }
-uselesskey-core-negative-pem = { workspace = true }
+uselesskey-core-hash = { workspace = true }
 
 [dev-dependencies]
 proptest.workspace = true
@@ -25,7 +24,7 @@ rstest.workspace = true
 
 [features]
 default = ["std"]
-std = ["uselesskey-core-negative-der/std", "uselesskey-core-negative-pem/std"]
+std = ["uselesskey-core-hash/std"]
 
 [package.metadata.docs.rs]
 features = ["std"]
